Fix failing spec when running on macOS with testing override but a broken libddprof
This is a bit of a corner case, but the outcome of tests on macOS
depends on all validations (like on Linux), we can't just assume
"macOS is OK" when `DD_PROFILING_MACOS_TESTING` is set.

Also moved the generic "operating system is not supported" test
outside of the test group to simplify testing.