Allow `HttpTransport._native_do_export` to be cancelled/interrupted
Up until now, `ddprof_ffi_ProfileExporterV3_send` was blocking and
nothing on the Ruby side could interrupt it.

This meant that if reporting a profile was slow/hung, then pressing
ctrl+c or calling `Thread#kill` would do nothing -- send would
only return after it hit the configured timeout (which by default is
30 seconds).

This was very annoying if we're talking about an app that is trying
to shut down and is being delayed by a report that is taking longer
than expected.

To fix this, we make use of the `CancellationToken` added in
<https://github.com/DataDog/libddprof/pull/45> to enable the Ruby VM
to signal to libddprof that it should cancel an in-progress `send`.

Thus, as the added spec shows, we can make `send` return early
by interrupting the thread, rather than needing to wait until the
timeout was hit.